Create Longest palindrome in a string
Time Complexity->O(|S|^2)
Space Complexity->O(1)
Approach->Brute force:- All the substrings are generated and each one is checked whether a palindrome or not and on comparing their length the longest one is returned.[O(N^3)]
Optimized approach:- Two pointers are taken and  they are expanded left and right until the characters at the two pointers are equal i.e , the substring is a palindrome (this is how this approach encompasses all the possible substrings without generating every one as in brute force) but two cases are considered one is odd case where the substring has odd no. of characters so only one pointer is taken so the left and right pointers are passed to the method as same index and second is even case where the substring has even no. of characters so the left and right pointers are passed as consecutive indexes and expand left and right respectively. When the equality at the left and right pointer indexes are becomes false then a substring starting from (left+1(since after the last iteration the left is incremented to -1 so it needs to be incremented to 0)->right) is returned as the palindromic substring.