Create Rearrange an array with O(1) extra space
Time Complexity->O(N)
Space Complexity->O(1)
Approach->The intuition  of without using a temporary storage is to devise a method of encoding the elements in such a way so that the encoded value can be operated by a common element(which is N since the array elements are less than N) and we can obtain both the old and transformed value in the array. Example- At index 0 , the old value is 5 and the new value is 4 and N=6 so (5+4)%N gives some value which has both 5 and 4 so we need to operate in such a way so that only one value is obtained as remainder while other fetches a remainder of 0.(5+N*4)%N will give 5 since 4*N is a multiple of N so it fetches 0 as remainder. But when we are encoding the value serially one by one then also we have to use certain values which have already been encoded so we need to use the decoding operation while encoding also, so the encoding operation becomes-(old value+N*(transformed value%N(Decoding operation))) and Decoding operation-Transformed value=encoded value/N.