Create Matchsticks game
Time Complexity->O(1)
Space Complexity->O(1)
Approach->It is a problem of game theory where we need to determine the winning states and loosing states. So, the winning states can be if the number of matches left are 1,2,3,4,6,7,8,9,11,.... Thus, non-multiples of 5 since maximum 4 matches can be chosen. Ex-If 1 or 2 or 3 or 4 matches are left then A can take all of them at once. If 6 or 7 or 8 or 9 matches are left then A can take 1 i.e, minimum matches then whatever B takes the states comes to 5 matches so the last match always left for A thus non-multiples of A are all winning states of A but for multiples for 5 , A can only take 4 so the last one always remains for B to take while for 10 or 15 , A will take a max of 4 matches then B will pick 1 match and the remaining 5 matches are left so it becomes a loosing state for A ,thus the multiples of 5 are loosing states for 5. If the number of matches is a multiple of 5 then A can never win so -1 is returned while if it is not then the modulus gives the answer which is the least number of matches greater than the lower closest multiple of 5.