Create Reverse Coding
Time Complexity->O(1)
Space Complexity->O(1)
Approach->It is a simple problem where we need to check that the integer values don't overflow. If we do this through iteration it will face a TLE since it is taking O(N) complexity so we have to do it in O(1) which is using the direct formula n*(n+1)/2 but there are possibilities of n,(n+1) or (n*(n+1)) to overflow so we need to perform modulus for each of these terms with 1000000007 or another method is to calculate the entire sum n*(n+1)/2 in a long type variable and lastly perform the modulus then explicitly converting to int the answer is returned.