Create Flip Bits
Time Complexity->O(N)
Space Complexity->O(1)
Approach-> It is a variation of the maximum sum subarray or Kadane's algorithm. So, after flipping a zero we get a one and on flipping a one we get a zero so in order to maximize the number of ones so in this context we can treat the ones as -1 since on flipping we are losing ones and we treat zeros as +1 since on flipping we get ones. We take another array where we assign +1 for zeros and -1 for ones and perform Kadane's algorithm on +1 and -1 array and we found the maximum subarray where we get the maximum sum so it is the subarray where we get maximum no. of ones but not necessarily the maximum sum in that subarray is equal to the number of ones formed after flipping all the elements in that subarray. So, the answer should be the maximum subarray sum plus the no. of ones in the original array.
This is because in the maximum subarray sum we have all the zeros that need to be flipped for getting the maximum ones in the entire array plus some ones in the array initially which falls in the subarray range which are getting flipped to zeros and negating the zeros flipped to ones in the subarray. And, the original no. of ones in the array also contains ones which are outside the maximum subarray range which are not getting affected by an flipping and the ones which are in the subarray range that are getting flipped to zeros .So, on adding the maximum subarray sum and the original no. of ones we get the answer since the effect of ones which are falling in the subarray range is getting neutralized and we get the no. of zeros in the subarray that needed to flipped to get the maximum no. of ones plus the ones which are outside the subarray that don't get affected due to flipping.