Create Prefix match with other strings
Time Complexity->O(N)
Space Complexity->O(1)
Approach-> If the length of the prefix is more than the reference string then we have to return 0 since the string has not many characters as per the prefix size so it can be never same so 0 is returned. If the length of the ref. string is more than the prefix length but the strings in the array are small as compared to the prefix length then also we are not counting those but if both the array string and ref. string are longer than prefix length then the prefixes are compared and if they are equal then the counter is incremented.