Create Minimum number of jumps
Time Complexity->O(N)
Space Complexity->O(1)
Intuition->For every current position , we can either make the longest jump which is equal to the value at that position or we can make a jump shorter than the value at that position. So, we need to chose either of the options to make sure so that no zeros come in between because we will get stuck and won't be able reach the end and secondly we need to make the minimum jumps which can be done by selecting the maximum jump values within the jump radius of the current position.
Approach->We need to check the maximum value within the jump radius of the current position and need to make that jump to that max value position so that we can complete the travel in minimum jumps. This is done by analyzing every position's value by taking the max of the current des variable and max jump radius of the current position and whenever pos becomes equal to i (i.e, the counter variable i reaches the shortest jumps path then we update the value of pos with des to steer and continue in that shortest jumps path only and keep on calculating the minimum jumps).Another caveat in the problem is if we encounter a zero at the beginning or in the middle where all the positions within the jump radius are leading to that zero position so we are stuck and won't proceed so we return the number of jumps within the loops itself because if we return the jumps outside it will be wrong if we get stuck in the middle due to zeros and if we do get stuck then the pos variable will never reach the end so the return statement if condition will never get executed and the loop will eventually run out and -1 is returned outside the loop since we got stuck.