Create Unique rows in Boolean matrix
Time Complexity->O(row*col)
Space Complexity->O(row*col)
Approach->We have taken a temporary arraylist and checking if it is not present in the arraylist of arraylists then only we are adding it. We are just treating every arraylist as a primitive date type and written the logic.