Create Union of Two Linked Lists
Time Complexity->O((N+M)*log(N+M))
Space Complexity->O(N+M)
Approach->In this problem we have to create a hashset which will only take unique elements then we need to copy it to a arraylist and then sort it. Then, we need to create a traversing node and keep on creating new nodes  and assign the next of the traversal node to the new node and the transfer the traversal node to the new node untill the end and assign the traversal node's next with null ending the linked list and returning the head node.