Create Happy Friends
Time Complexity->O(N)
Space Complexity->O(1)
Approach->In this problem two arrays are present where prices array representing the prices of different dishes and the dislike array which indicates that every dish is disliked by at least one person so every index is represented by a dish and the value at a particular index stated the person who disliked that particular dish. So we are required to find the minimum cost to buy dishes where every dish can be shared by all persons. So , in order to minimize the cost we need to find the minimum cost dish so it will have at least one person who dislikes it so we shall find the second lowest priced dish which is not disliked by the person who disliked the east priced dish .Another case is when the person who dislikes the least priced dish dislikes all the other dishes so we won't be able to buy any combination of dishes so we have to return -1.