Create nCr
Time Complexity->O(N*R)
Space Complexity->O(R)
Approach->For finding nCr we have to generate the pascal's triangle where n is the no. of rows of the triangle and since nCr and nC(n-r) have equal values so in the nth row at rth or (n-r)th position depending whichever is minimum we will get the desired answer. So, we will create an array where and for each row(at each iteration of the outer loop) the respective row of the pascal's triangle will be generated from the previous row. The ans will be found at the minimum of rth and (n-r)th position so the dp array is found at min(r,n-r) the position. Every time the array's value is being changed since at every iteration that level of the pascal's triangle is being generated in the array. The value at any position in the pascal's triangle is the sum of the value at that position in the previous row and the value at the position just before in the previous row. So, the inner loop is run from r to 1  since at every row the 0th index value is always 1.