Create Is Binary Number Multiple of 3
Time Complexity->O(N) ; N is length of the string
Space Complexity->O(1)
Approach->If we try the conventional approach then for very large number, a large power of 2 is difficult to be calculated so we need to try some other method. The second approach included all the constituent numbers in binary format to be taken out. For every constituent numbers it can be broken down into a part which is divisible by 3(N-N%3) and a part which is not divisible by 3(N%3).If we add all the divisible parts then definitely it is divisible by 3 but for the other part may or may not be divisible by 3 which decides the number overall is divisible by 3 or not. So by aptitude remainder theorem if the power of 2 is odd then remainder or the part not divisible by 3 in the number is 2 and if the power is even then the remainder is 1.So we add all these remainders and if the sum is divisible by 3 then we return 1 or else we return 0.