Create Maximum Index
Time Complexity->O(N)
Space Complexity->O(N)
Intuition->If we have to do a straightforward O(N^2) problem in O(N) T.C then we have to use any one of the approaches-two pointer, sliding window, greedy, etc. So, in this problem two pointer approach is used so the two pointer approach always involves the idea of eliminating the redundant iterations based on observation. 
Approach->We have to create two arrays which are a prefix array in which at every index the minimum of all the elements from the left end up to that index is stored at that index while a suffix array in which at every index the largest of all the elements from the right end up to that index is stored at that index. These arrays are created to reduce the redundant iterations since if one index in the prefix array satisfies the condition  arr[j]>=arr[i] then rest of the indexes in the prefix array is not checked as they are present after that index and whether they satisfy the condition or not doesn't matter as the difference of the indexes has to be maximized so the leftmost condition satisfying index is only considered and if in between the condition becomes false then there is a possibility that some index in the prefix array after the current satisfying index may satisfy the condition but element are not considered from the first again since the difference in indices will always be less so the index in the suffix array where the condition failed serves as the starting point for the next comparison cycle and if the index in the prefix array is truly the answer then the pointer in the suffix array will eventually run out and we will get the answer otherwise the condition again fails and the pointer in the prefix array is again incremented indicating another possibility of an answer.