Create Count the subarrays having product less than k
Time Complexity->O(N)
Space Complexity->O(1)
Intuition->The brute force approach is based on nested loops but in order to optimize it as always we need to reduce the redundant iterations for which we can use approaches like two pointers and sliding window. So, for linear complexity and traversal of the array only once at every index we need to observe that if the cumulative product up to that index is greater than k then we don't need to consider  the subarrays beyond that index thus optimizing the code. And another caviet is, if up to a certain index we get the cumulative product greater than k then there might be a possibility that some subarrays within that index may satisfy the condition so that needs to be taken into account and for this we need to remove the elements from the front until the product again becomes less than k so we will again start considering from that point when the cumulative product is less than k and again we start counting from this point cause at any previous indexes the cumulative product is more than k so there won't be any subarrays satisfying the condition.
Approach-> We use the sliding window approach where we take 2 pointers and constantly shift the right index and at every point we get the cumulative sum to check whether it is less than k or not then if it is then we calculate the number of subarrays in the current window which is calculated by (right-left)+1 and add it to our answer. But if is more than k then we don't shift the right pointer as multiplying any number more the product will always be greater than k so there is no point in considering so we shift the left pointer cause there might be a possibilty that starting from some left index the cumulative product is less than k indicating there might be some internal subarrays so we go on removing the elements from the left and shift the left pointers until the product again becomes less than k and start shifting the right pointer again by keeping the left pointer fixed.